refactor(api): stay within vercel hobby function limits
