--- api/chat/push-config.js
@@ -1,5 +1,0 @@
-import { createChatPushConfigHandler } from '../../src/chat/push-api.js';
-
-const handler = createChatPushConfigHandler();
-
-export default handler;
