--- api/chat/push-notify.js
@@ -1,5 +1,0 @@
-import { createChatPushNotifyHandler } from '../../src/chat/push-api.js';
-
-const handler = createChatPushNotifyHandler();
-
-export default handler;
