--- api/chat/push-subscriptions.js
@@ -1,5 +1,0 @@
-import { createChatPushSubscriptionHandler } from '../../src/chat/push-api.js';
-
-const handler = createChatPushSubscriptionHandler();
-
-export default handler;
