+++ api/chat/push.js
@@ -1,0 +1,5 @@
+import { createChatPushHandler } from '../../src/chat/push-api.js';
+
+const handler = createChatPushHandler();
+
+export default handler;
