+++ chat/index.html
@@ -318,9 +318,9 @@ const notificationsPreferenceKey = 'chatNotificationsEnabled';
 const notificationIconPath = '/icons/icon-192.png';
 const notificationStateStorageKey = 'chatNotificationState';
 const pushSubscriptionIdStorageKey = 'chatPushSubscriptionId';
-const pushConfigEndpoint = '/api/chat/push-config';
-const pushSubscriptionEndpoint = '/api/chat/push-subscriptions';
-const pushNotifyEndpoint = '/api/chat/push-notify';
+const pushConfigEndpoint = '/api/chat/push';
+const pushSubscriptionEndpoint = '/api/chat/push';
+const pushNotifyEndpoint = '/api/chat/push?action=notify';
 const INITIAL_SYNC_GRACE_MS = 1500;
 let notificationsEnabled = false;
 let serviceWorkerRegistration = null;
